begin refactoring program to run through ScreenManager class
removed console UI from Screens and renamed to ScreenManager
